pyproject.toml
[tool.poetry]
name = "reksoft"
version = "0.1.0"
description = ""
authors = ["antoshayastreb <nedanon@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
waitress = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
